committing on new branch - all methods marked for deletion deleted and versioning added
