+++ Python_Engine/Python_Engine.csproj
@@ -67,11 +67,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Compute\InstallBasePythonEnvironment.cs" />
-    <Compile Include="Compute\InstallReferencedVirtualenv.cs" />
-    <Compile Include="Compute\InstallVirtualenv.cs" />
-    <Compile Include="Compute\Logging.cs" />
-    <Compile Include="Compute\RemoveVirtualenv.cs" />
     <Compile Include="Compute\Run.cs" />
     <Compile Include="Compute\BasePythonEnvironment.cs" />
     <Compile Include="Compute\Remove.cs" />
@@ -80,19 +75,9 @@
     <Compile Include="Compute\VirtualEnvironment.cs" />
     <Compile Include="Compute\UnzipFile.cs" />
     <Compile Include="Compute\Download.cs" />
-    <Compile Include="Compute\Invoke.cs" />
-    <Compile Include="Convert\ToPython.cs" />
     <Compile Include="Modify\AddQuotesIfRequired.cs" />
-    <Compile Include="Query\CodeDirectory.cs" />
-    <Compile Include="Query\EnvironmentExists.cs" />
-    <Compile Include="Query\EnvironmentsDirectory.cs" />
-    <Compile Include="Query\ExtensionsDirectory.cs" />
-    <Compile Include="Query\ToString.cs" />
-    <Compile Include="Query\VirtualEnv.cs" />
-    <Compile Include="Query\VirtualEnvPythonExePath.cs" />
-    <Compile Include="Query\VirtualEnvDirectory.cs" />
     <Compile Include="Query\VirtualEnvironment.cs" />
-    <Compile Include="Query\Version.cs" />
+    <Compile Include="Query\PythonVersion.cs" />
     <Compile Include="Query\Directory.cs" />
     <Compile Include="Query\IsValidEnvironmentName.cs" />
     <Compile Include="Query\EmbeddableURL.cs" />
@@ -111,6 +96,7 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
+    <Folder Include="Convert\" />
     <Folder Include="Create\" />
   </ItemGroup>
   <ItemGroup>
