--- Python_Engine/Query/ToString.cs
@@ -1,42 +1,0 @@
-﻿/*
- * This file is part of the Buildings and Habitats object Model (BHoM)
- * Copyright (c) 2015 - 2023, the respective contributors. All rights reserved.
- *
- * Each contributor holds copyright over their respective contributions.
- * The project versioning (Git) records all such contribution source information.
- *
- *
- * The BHoM is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3.0 of the License, or
- * (at your option) any later version.
- *
- * The BHoM is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.
- */
-
-using BH.oM.Base.Attributes;
-
-using System.ComponentModel;
-using System.IO;
-
-namespace BH.Engine.Python
-{
-    public static partial class Query
-    {
-        // TODO - REMOVE THIS METHOD, NO LONGER REQUIRED
-        [Description("The string representation of a BHoM PythonEnvironment object.")]
-        [Input("env", "A BHoM PythonEnvironment object.")]
-        [Output("The string representation of the input BHoM PythonEnvironment object.")]
-        public static string ToString(this oM.Python.PythonEnvironment env)
-        {
-            return $"{env.GetType().Name}[{env.Name}]";
-        }
-    }
-}
-
