--- Python_Engine/Query/VirtualEnvDirectory.cs
@@ -1,51 +1,0 @@
-﻿/*
- * This file is part of the Buildings and Habitats object Model (BHoM)
- * Copyright (c) 2015 - 2023, the respective contributors. All rights reserved.
- *
- * Each contributor holds copyright over their respective contributions.
- * The project versioning (Git) records all such contribution source information.
- *
- *
- * The BHoM is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3.0 of the License, or
- * (at your option) any later version.
- *
- * The BHoM is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.
- */
-
-using BH.oM.Base.Attributes;
-
-using System.ComponentModel;
-using System.IO;
-using System.Xml.Linq;
-
-namespace BH.Engine.Python
-{
-    public static partial class Query
-    {
-        // TODO - REMOVE THIS METHOD, NO LONGER REQUIRED
-        [Description("The location where Python.exe for a named virtual environment would reside. This method does not check if that environment actually exists or not.")]
-        [Input("envName", "The virtual environment name.")]
-        [Output("The location where a Python virtual environment would reside.")]
-        public static string VirtualEnvPythonExePath(string envName)
-        {
-            string exePath = Path.Combine(VirtualEnvDirectory(envName), "Scripts", "python.exe");
-
-            if (File.Exists(exePath))
-            {
-                return exePath;
-            }
-
-            BH.Engine.Base.Compute.RecordError($"The executable for \"{envName}\" does not exist at \"{exePath}\".");
-            return null;
-        }
-    }
-}
-
