+++ Python_Engine/Versioning_63.json
@@ -1,0 +1,20 @@
+﻿{
+  "MessageForDeleted": {
+		"BH.Engine.Python.Compute.InstallBasePythonEnvironment(System.Boolean)": "This method has been replaced. Please use the BasePythonEnvironment method instead.",
+		"BH.Engine.Python.Compute.InstallReferencedVirtualenv(System.String, System.String, System.String, System.Boolean)": "This method has been removed.",
+		"BH.Engine.Python.Compute.InstallVirtualenv(System.String, BH.oM.Python.Enums.PythonVersion, System.String, System.Boolean)": "This method has been removed.",
+		"BH.Engine.Python.Compute.LoggingHeader(System.String)": "This method has been removed.",
+		"BH.Engine.Python.Compute.RemoveVirtualenv(System.String, System.Boolean)": "This method has been removed.",
+		"BH.Engine.Python.Query.CodeDirectory()": "This method has been removed.",
+		"BH.Engine.Python.Query.EnvironmentExists(BH.oM.Python.PythonEnvironment)": "This method has been removed.",
+		"BH.Engine.Python.Query.ExtensionsDirectory()": "This method has been removed.",
+		"BH.Engine.Python.Query.ToString(BH.oM.Python.PythonEnvironment)": "This method has been removed.",
+		"BH.Engine.Python.Query.Virtualenv(System.String)": "This method has been removed.",
+		"BH.Engine.Python.Query.VirtualEnvPythonExePath(System.String)": "This method has been removed.",
+		"BH.Engine.Python.Query.VirtualEnvDirectory(System.String)": "This method has been removed.",
+		"BH.Engine.Python.Convert.ToPython(System.Array(,))": "This method has been removed.",
+		"BH.Engine.Python.Compute.RunCommandAsync(System.String, System.Boolean, System.String)": "This method has been removed.",
+		"BH.Engine.Python.Compute.RunCommandBool(System.String, System.Boolean, System.String, System.Double)": "This method has been removed.",
+		"BH.Engine.Python.Compute.RunPythonScript(BH.oM.Python.PythonEnvironment, System.String, System.List(String))": "This method has been removed.",
+	}
+}
